requirements.txt
requests
beautifulsoup4
python-dateutil
tenacity
python-dotenv
sqlalchemy
